backend/requirements.txt
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
boto3==1.38.27
botocore==1.38.27
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
envparse==0.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
sib-api-v3-sdk==7.6.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
passlib==1.7.4
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
ruff==0.12.4
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.2
yarl==1.20.1